🚸 Dans les paramètre si les valeur sont entière elles sont affiché comme tel

+++ app/src/main/java/com/fsp/ticket_resto_app/controlleur/Controlleur.java
@@ -48,11 +48,11 @@ public class Controlleur {
         return decomposition.decomposeReste();
     }
 
-    public void setValeurTicket1(int valeur) {
+    public void setValeurTicket1(double valeur) {
         decomposition.setPremiereValeur(valeur);
     }
 
-    public void setValeurTicket2(int valeur) {
+    public void setValeurTicket2(double valeur) {
         decomposition.setDeuxiemeValeur(valeur);
     }
